+++ src/garage/np/policies/__init__.py
@@ -4,4 +4,8 @@ from garage.np.policies.fixed_policy import FixedPolicy
 from garage.np.policies.policy import Policy
 from garage.np.policies.scripted_policy import ScriptedPolicy
 
-__all__ = ['FixedPolicy', 'Policy', 'ScriptedPolicy']
+__all__ = [
+    'FixedPolicy',
+    'Policy',
+    'ScriptedPolicy',
+]